clean up and reorganize codebase
