--- biencoders/dataset/sample_audiocaps.csv
@@ -1,3 +1,0 @@
-file_name,caption
-sample1.wav,A dog barking
-sample2.wav,People talking
